requirements.txt
numpy==2.3.1
pandas==2.3.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
distro==1.9.0
et-xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.97.0
openpyxl==3.1.5
pydantic==2.11.7
pydantic-core==2.33.2
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
click==8.2.1
fastapi==0.116.1
starlette==0.47.1
uvicorn==0.35.0
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
packaging==25.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
tokenizers==0.21.2
transformers==4.53.2
urllib3==2.5.0
python-dotenv==1.1.1

jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
sympy==1.14.0

accelerate==1.10.0
psutil==7.0.0
pillow==11.3.0

nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
pyjwt==2.10.1
protobuf==5.27.2

fastapi
streamlit
pymupdf
frontend
sentence_transformers
pandas
pyarrow
dill
aiohttp
numpy
accelerate